Place button row
